requirements.txt
aiohttp==3.11.11
asyncio==3.4.3
cloudscraper==1.2.71
colorama==0.4.6
fake-useragent==2.0.3
loguru==0.7.0
python-dotenv==1.0.1
requests==2.32.3
